Refactoring Blender Code
This is to support automated testing later own down the development line
- Added MMD Translation file into the code to allow for Translation of
	MMD models
