+++ hifi_tools/armature/__init__.py
@@ -1,0 +1,13 @@
+if "bpy" in locals():
+    import importlib
+    
+    importlib.reload(panel)
+    importlib.reload(repose)
+
+else:
+    from . import (
+        panel,
+        repose
+    )
+
+    import bpy
