+++ hifi_tools/files/__init__.py
@@ -1,0 +1,8 @@
+if "bpy" in locals():
+    import importlib
+    importlib.reload(hifi_json)
+else:
+    from . import (
+        hifi_json
+    )
+    import bpy
