+++ hifi_tools/utils/__init__.py
@@ -1,0 +1,8 @@
+if "bpy" in locals():
+    import importlib
+    importlib.reload(helpers)
+else:
+    from . import (
+        helpers
+    )
+    import bpy
