+++ hifi_tools/world/__init__.py
@@ -1,0 +1,12 @@
+if "bpy" in locals():
+    import importlib
+    importlib.reload(primitives)
+    importlib.reload(material)
+    importlib.reload(scene)
+else:
+    from . import (
+        primitives,
+        material,
+        scene
+    )
+    import bpy
