+++ hifi_tools/world/primitives.py
@@ -21,7 +21,6 @@
 
 
 import bpy
-from .hifi_scene_import import *
 
 import addon_utils
